requirements.txt
django
django-kronos
django-celery
django-filter
djangorestframework
markdown
